remove unused code
Note take `length_object` for example, it is unknown, we can remove and inline it
