--- ocaml-tree/iter_maker.js
@@ -1,157 +1,0 @@
-//@ts-check
-var assert = require("assert");
-var node_types = require("./node_types");
-var init = node_types.init;
-/**
- *
- * @typedef {import('./node_types').Node} Node
- */
-
-/**
- *
- * @param {{name:string, def:Node}} typedef
- * @param {Set<string>} allNames
- * @returns {string}
- */
-function mkMethod({ name, def }, allNames) {
-  return `method ${name} : ${name} -> unit = ${mkBody(def, allNames)}  `;
-}
-
-var skip = `ignore`;
-
-/**
- * @param {Node} def
- * @param {Set<string>} allNames
- */
-function mkBody(def, allNames) {
-  // @ts-ignore
-  assert(def !== undefined);
-  switch (def.type) {
-    case "type_constructor_path":
-      var basic = node_types.isSupported(def, allNames);
-      if (basic !== undefined) {
-        return `_self#${basic}`;
-      }
-      return skip;
-    case "constructed_type":
-      // FIXME
-      var [list, base] = [...def.children].reverse();
-
-      switch (list.text) {
-        case "option":
-        case "list":  
-          var inner = mkBody(base, allNames);
-          if (inner === skip) {
-            return inner;
-          }
-          return `(${list.text} ${inner})`;
-        // case "list":
-        //   // there are list and other
-        //   return `(${mkBody(list, allNames)} ${mkBody(base, allNames)})`;
-        default:
-          throw new Error(`not supported high order types ${list.text}`);
-      }
-
-    case "record_declaration":
-      var len = def.children.length;
-      var args = init(len, (i) => `_x${i}`);
-      var pat_exp = init(len, (i) => {
-        return `${def.children[i].mainText} = ${args[i]}`;
-      });
-
-      /**
-       * @type {string[]}
-       */
-      var body = args
-        .map((x, i) => {
-          var ty = def.children[i].children[1];
-          return mkBodyApply(ty, allNames, x);
-        })
-        .filter(Boolean);
-      return `fun { ${pat_exp.join(";")}} -> begin ${body.join(";")} end`;
-    case "variant_declaration":
-      var len = def.children.length;
-      var branches = def.children.map((branch) => mkBranch(branch, allNames));
-      return `function \n| ${branches.join("\n|")}`;
-    case "tuple_type":
-      var len = def.children.length;
-      var args = init(len, (i) => `_x${i}`);
-      var body = args
-        .map((x, i) => mkBodyApply(def.children[i], allNames, x))
-        .filter(Boolean);
-      return `(fun ( ${args.join(",")}) -> begin ${body.join(";")} end)`;
-    default:
-      throw new Error(`unkonwn ${def.type}`);
-  }
-}
-
-/**
- *
- * @param {Node} ty
- * @param {Set<string>} allNames
- * @param {string} arg
- */
-function mkBodyApply(ty, allNames, arg) {
-  var fn = mkBody(ty, allNames);
-  if (fn === skip) {
-    return ``;
-  }
-  return `${fn} ${arg}`;
-}
-
-/**
- *
- * @param {Node} branch
- * branch is constructor_declaration
- * @param {Set<string>} allNames
- * @returns {string}
- */
-function mkBranch(branch, allNames) {
-  // @ts-ignore
-  assert(branch?.type === "constructor_declaration");
-  var [{ text }, ...rest] = branch.children;
-  // TODO: add inline record support
-  var len = rest.length;
-  if (len === 0) {
-    return `${text} -> ()`;
-  }
-  var args = init(len, (i) => `_x${i}`);
-  var pat_exp = `${text} ( ${args.join(",")}) `;
-  var body = args
-    .map((x, i) => {
-      var ty = rest[i];
-      return mkBodyApply(ty, allNames, x);
-    })
-    .filter(Boolean);
-  if (body.length === 0) {
-    return `${text} _ -> ()`;
-  }
-  return `${pat_exp} -> \n begin ${body.join(";")} end`;
-}
-
-/**
- *
- * @param {{name : string, def: Node}[]} typedefs
- * @returns {string}
- */
-function make(typedefs) {
-  var allNames = new Set([...typedefs.map((x) => x.name), "option", "list"]);
-  var output = typedefs.map((x) => mkMethod(x, allNames));
-  var o = `
-    open J  
-
-    let option sub  v =
-      match v with 
-      | None -> ()
-      | Some v -> sub  v
-    let rec list sub v =
-      match v with 
-      | [] -> ()
-      | x::xs -> sub x ; list sub xs 
-    class iter = object (_self : 'self_type)
-    ${output.join("\n")}    
-    end
-    `;
-  return o;
-}
-exports.make = make;
