+++ ocaml-tree/types.d.ts
@@ -1,0 +1,7 @@
+
+
+export interface Obj {
+    eta : string
+    beta : (x : string) => string  
+    method? : string
+}
