+++ ocaml-tree/wasm.js
@@ -10,14 +10,8 @@ var mode = "map";
 for (let i = 0; i < process.argv.length; ++i) {
   let u = process.argv[i];
   switch (u) {
-    case "-map":
-      mode = "map";
-      break;
     case "-fold":
       mode = "fold";
-      break;
-    case "-iter":
-      mode = "iter";
       break;
     case "-record-iter":
       mode = "record-iter";
@@ -37,9 +31,8 @@ for (let i = 0; i < process.argv.length; ++i) {
 }
 var source = fs.readFileSync(input, "utf8");
 var node_types = require("./node_types");
-var map_maker = require("./map_maker");
+
 var fold_maker = require("./fold_maker");
-var iter_maker = require("./iter_maker");
 var record_iter = require("./record_iter");
 var record_map = require("./record_map");
 
@@ -55,14 +48,8 @@ var maker = node_types.maker;
   var out = p.parse(source);
   var typedefs = node_types.getTypedefs(out);
   switch (mode) {
-    case "map":
-      fs.writeFileSync(output, maker(map_maker.make, typedefs), "utf8");
-      break;
     case "fold":
       fs.writeFileSync(output, maker(fold_maker.make, typedefs), "utf8");
-      break;
-    case "iter":
-      fs.writeFileSync(output, maker(iter_maker.make, typedefs), "utf8");
       break;
     case "record-iter":
       fs.writeFileSync(output, maker(record_iter.make, typedefs), "utf8");
